feat(import): add safe assisted work data ingestion

+++ backend/src/main/java/com/alveryn/api/workrecord/dto/WorkRecordLineRequest.java
@@ -16,5 +16,5 @@ public record WorkRecordLineRequest(
     LocalTime endTime,
     @PositiveOrZero Integer durationMinutes,
     @PositiveOrZero Integer unpaidBreakMinutes,
-    @PositiveOrZero @Max(1000) Integer extraPayPercentage,
+    @PositiveOrZero @Max(1000) BigDecimal extraPayPercentage,
     @Size(max = 500) String notes) {}
