+++ backend/src/main/java/com/alveryn/api/workrecord/line/dto/WorkRecordLineResponse.java
@@ -33,5 +33,5 @@ public record WorkRecordLineResponse(
     BigDecimal baseGrossAmount,
     BigDecimal extraGrossAmount,
     BigDecimal totalGrossAmount,
-    int extraPayPercentage,
+    BigDecimal extraPayPercentage,
     String notes) {}
